docs(react): update nav-menu sandboxes
