+++ docs/app/react/nav-menu/components/nav-menu-preview.tsx
@@ -1,0 +1,17 @@
+import {
+  NavMenu,
+  NavMenuLink,
+  NavMenuList,
+  NavMenuTrigger,
+} from '@cerberus-design/react'
+
+export function BasicNavMenuPreview() {
+  return (
+    <NavMenu>
+      <NavMenuTrigger controls="basic:nav">Features</NavMenuTrigger>
+      <NavMenuList id="basic:nav">
+        <NavMenuLink>Something</NavMenuLink>
+      </NavMenuList>
+    </NavMenu>
+  )
+}
