+++ packages/react/src/index.ts
@@ -3,6 +3,7 @@
 export * from './components/Button'
 export * from './components/NavMenuTrigger'
 export * from './components/NavMenuList'
+export * from './components/NavMenuLink'
 export * from './components/Show'
 
 // context
